reformatted + refactored
- lower-case naming convention
- removed unused code
- reformatted the newly added reactive code
