--- src/components/Editor/index.tsx
@@ -1,26 +1,0 @@
-import * as React from 'react';
-import { Module } from '../../syntax-tree/module';
-
-import './style.css'
-
-
-type Props = {className}
-type State = {}
-
-
-
-class EditorComponent extends React.Component<Props, State>{
-    static EditorParentId = "mainEditorDiv"
-  
-    render(){
-      return (
-        <div id = {EditorComponent.EditorParentId} className = {this.props.className}>
-        </div>
-    );
-  }
-}
-
-export default EditorComponent;
-
-
-
