--- src/editor/visuals.ts
@@ -1,17 +1,0 @@
-/*import { CodeConstruct } from "../syntax-tree/ast";
-
-
-export class Visuals{
-
-    constructor(){}
-
-
-    addDraftModeHighlight(code: CodeConstruct){
-        const highlightElement = 
-    }
-}
-*/
-
-
-
-
